Add generator producing MLIR expressions
The generator can be invoked with the -emit=mlir option of the teckyl
binary. For each tensor kernel, this generates an MLIR FuncOp with the
same name, whose body is composed of a set of affine loop nests.

In contrast to TC, iterators and their domains are entirely local to
an assignment / a reduction. Also, the generator currently does not
support any inference. Iteration domains for all iterators must be
explicitly stated with where clauses for every expression and the
shape and type of output parameters must be provided in the signature.