Add support for type suffixes in number literals
Allow the type of a number literal to be specified as a type
suffix. The following suffixes are supported:

  - u8: unsigned, 8-bit integer
  - u16: unsigned, 16-bit integer
  - u32: unsigned, 32-bit integer
  - u64: unsigned, 64-bit integer

  - i8: signed, 8-bit integer
  - i16: signed, 16-bit integer
  - i32: signed, 32-bit integer
  - i64: signed, 64-bit integer

  - f16: 16-bit IEEE floating point value
  - f32: 32-bit IEEE floating point value
  - f64: 64-bit IEEE floating point value

Literals without a type suffix are treated as before: integer literals
without suffix are treated as signed 32-bit integers and floating
point literals without suffix are treated as 32-bit IEEE floating
point values.