Implicitly convert types of operands if conversion is lossless
Teckyl conservatively aborts if the types of the operands of a binary
operation do not match or if the type of a value assigned to a tensor
element does not match the element's type.

This patch adds implicit conversions if the types in the
above-mentioned situations do not match, but only if the conversion is
lossless (i.e., if all values of the source type can be represented by
the target type). The following rules apply:

  - A signed integer can be converted to a wider signed integer (e.g.,
    int8 to int32)

  - A floating point value can be converted to a wider floating point
    value (e.g., float16 to float32)

  - An integer can be converted to a floating point value if the
    mantissa is at least as wide as the integer type (e.g., an int16
    can be converted to a float32, whose mantissa has 23 bits)

Conversions from and to unsigned integers are not yet supported, as
Teckyl currently only handles signed integers.