Explicitly convert loop bounds to Index values
LoopOps require loop bounds to be Index values. This patch adds an
implicit conversion for loop bounds that are not Index values. Since
the bit width of Index is platform dependent and cannot be queried
from the IR, no checks can be performed prior to the conversion and
converted values might silently be truncated.