Properly check conditions for linalg.generic in code generation for reductions
Until now, the code generator only checks for affine tensor indexing
as a prerequisite for generating linalg.generic operations. This
completely ignores the bounds specified in where clauses and doesn't
check if the generated linalg code can be lowered subsequently.

This patch adds additional checks before generating a linalg.generic
operation beyond checking for affine tensor indexing:

  - A check for the existence of a direct mapping between iteration
    dimensions and tensor accesses as required by linalg.generic. This
    requires that each iterator of the comprehension is referenced at
    least once for direct indexing.

  - A check ensuring that the bounds for the comprehension for
    iterators with direct indexing match the size of the indexed
    tensor dimension.

As a by-product, the checks enforce the presence of where clauses for
each reduction and properly take into account the bounds.

Furthermore, for simple tensor initializations for default-intialized
reductions, a view with shifted offsets and adjusted sizes to restore
the conditions for linalg.generic is created if necessary.