Add tool for testing transformations of arithmetic expressions
This patch adds a new tool `transform`, capable of performing
transformations on arithmetic expressions for range inference. The
tool takes as input an arithmetic expression that is parsed using the
tool's expression parser, performs a transformation specified as a
command line parameter and outputs the transformed expression.

Example usage:

  `echo '2*(i+j)' | transform -trafo=distr`

distributes the multiplication of the expression `2*(i+j)` on the sum
and yields:

  `((2*i)+(2*j))`

The tool is intended for use in automated tests of transformations on
expressions used in range inference.

Co-authored-by: Andi Drebes <andi@drebesium.org>