BUGFIX: Cast static iterator bounds to index
The arithmetic in buildTensorInitialization() assumes that iterator
bounds are always of type index and unconditionally uses IndexCastOp
for conversion to 64-bit integers. However, static iterator bounds are
instantiated as integer values, which causes the
buildTensorInitialization to create an IndexCastOp from an integer
value to a 64-bit integer, resulting in a module verification error.

This patch converts the bounds to the index type prior to the
conversion to 64-bit integers if necessary.