Generalize definitions in for memrefs with macros in tests/exec/lib/memref.h
The header file tests/exec/lib/memref.h currently defines structs and
functions only for 2d float memrefs. This patch introduces the four
macros

  DECL_VEC1D_STRUCT,
  DECL_VEC2D_STRUCT,
  DECL_VEC1D_FUNCTIONS, and
  DECL_VEC2D_FUNCTIONS

that allow for the definition of one and two-dimensional memrefs with
arbitrary element types and axuiliary functions for these types. The
existing definitions for vec_f2d have been replaced with invocations
of the macros above.