Abort if input file cannot be opened
The function readFile() silently fails for files which cannot be
opened and simply returns an empty string. This prevents callers from
distinguishing between empty files and I/O errors. This patch causes
readFile() to throw an exception (if compiled with exceptions) or to
abort if the input file cannot be opened.