Teckyl: Add option -emit=header generating header file with function signatures
The option generates a C99 header file with the signatures for all
functions of the source file and wrapper functions that take bare
pointers and size parameters for parametric dimensions instead of
memref arguments.

E.g., for the following input, definition

  def mm(float(M,128) A, float(128,N) B) -> (float(M,N) C) { ... }

this generates the signature:

  void mm(const float* A_allocatedPtr,
	  const float* A_alignedPtr,
	  int64_t A_offset,
	  int64_t A_size0,
	  int64_t A_size1,
	  int64_t A_stride0,
	  int64_t A_stride1,

	  const float* B_allocatedPtr,
	  const float* B_alignedPtr,
	  int64_t B_offset,
	  int64_t B_size0,
	  int64_t B_size1,
	  int64_t B_stride0,
	  int64_t B_stride1,

	  float* C_allocatedPtr,
	  float* C_alignedPtr,
	  int64_t C_offset,
	  int64_t C_size0,
	  int64_t C_size1,
	  int64_t C_stride0,
	  int64_t C_stride1)

an a wrapper function with the following signature:

  static inline void
  mm_wrap(const float* A,
          const float* B,
	  float* C,
	  uint64_t M, uint64_t N)

A preprocessor symbol for the inclusion guard protecting against
double inclusion can be specified with the option -include-guard.